% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculatescITDG.R
\name{CalculatescITDG}
\alias{CalculatescITDG}
\title{Title}
\usage{
CalculatescITDG(
  object,
  celltype.list,
  p2p.deg.use,
  logfc.threshold = 0.25,
  pvalue.threshold = 0.05,
  celltype.use,
  time.points.use,
  time.points,
  sample.ncell.use = 200,
  seed.use = NULL,
  n.cores = 1,
  save.wd = NULL
)
}
\arguments{
\item{save.wd}{}
}
\description{
Title
}
