% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calFitModel.R
\name{calFitModel}
\alias{calFitModel}
\title{Title}
\usage{
calFitModel(
  x,
  modelFormulaStr,
  expressionFamily,
  relative_expr,
  disp_func = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
