% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disp_calc_helper_NB.R
\name{disp_calc_helper_NB}
\alias{disp_calc_helper_NB}
\title{Title}
\usage{
disp_calc_helper_NB(ITDGds, expressionFamily, min_cells_detected)
}
\arguments{
\item{min_cells_detected}{}
}
\description{
Title
}
