% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateDispersionsForMatrix.R
\name{estimateDispersionsForMatrix}
\alias{estimateDispersionsForMatrix}
\title{Title}
\usage{
estimateDispersionsForMatrix(
  ITDGds,
  modelFormulaStr = "~ 1",
  relative_expr = TRUE,
  min_cells_detected = 1,
  remove_outliers = TRUE
)
}
\arguments{
\item{remove_outliers}{}
}
\description{
Title
}
