% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFitModel.R
\name{getFitModel}
\alias{getFitModel}
\title{Title}
\usage{
getFitModel(
  x,
  trend_formula,
  expressionFamily,
  relative_expr,
  new_data,
  disp_func
)
}
\arguments{
\item{disp_func}{}
}
\description{
Title
}
