% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getP2PDEGs.R
\name{getP2PDEGs}
\alias{getP2PDEGs}
\title{Title}
\usage{
getP2PDEGs(
  object,
  celltype.use = NULL,
  celltype = NULL,
  time.points.use = NULL,
  time.points = NULL,
  cellnum.min = 10,
  logfc.threshold = 0.25,
  test.use = "wilcox"
)
}
\arguments{
\item{test.use}{}
}
\description{
Title
}
