% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HighFreqGene.R
\name{HighFreqGene}
\alias{HighFreqGene}
\title{Title}
\usage{
HighFreqGene(
  deg.list,
  moduleID,
  top.num = 50,
  y_l = -2,
  text.size = 2,
  anno.cols = NULL
)
}
\arguments{
\item{deg.list}{}

\item{moduleID}{}

\item{top.num}{}

\item{y_l}{}

\item{text.size}{}

\item{anno.cols}{}
}
\description{
Title
}
