% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametricDispersionFit.R
\name{parametricDispersionFit}
\alias{parametricDispersionFit}
\title{Title}
\usage{
parametricDispersionFit(disp_table, initial_coefs = c(1e-06, 1))
}
\arguments{
\item{initial_coefs}{}
}
\description{
Title
}
