% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessscITDG.R
\name{ProcessscITDG}
\alias{ProcessscITDG}
\title{Title}
\usage{
ProcessscITDG(
  object,
  celltype.use,
  p2p.deg.list,
  cal.celltype,
  time.points.use,
  time.points,
  sample.ncell.use,
  seed.use,
  GeneExpCur.wd
)
}
\arguments{
\item{GeneExpCur.wd}{}
}
\description{
Title
}
