% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scITDGPlot.R
\name{scITDGPlot}
\alias{scITDGPlot}
\title{Title}
\usage{
scITDGPlot(
  object,
  k.num = 6,
  cluster.order.list = NULL,
  show.trajectory = TRUE,
  show.term = TRUE,
  OrgDb = org.Hs.eg.db,
  save.wd = NULL,
  use_raster = TRUE,
  plot.height = NULL,
  heatmap.cols = NULL,
  scale.adj = NULL
)
}
\arguments{
\item{scale.adj}{}
}
\description{
Title
}
