% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scITDGIntegratePlot.R
\name{scITDGIntegratePlot}
\alias{scITDGIntegratePlot}
\title{Title}
\usage{
scITDGIntegratePlot(
  exp.cur,
  row.dist = row.dist,
  k.num = 15,
  time.points = time.points,
  sample.ncell.use = sample.ncell.use,
  integrate.cluster.list = NULL,
  cluster.order.list = NULL,
  show.trajectory = TRUE,
  show.term = TRUE,
  OrgDb = org.Hs.eg.db,
  save.wd = NULL,
  use_raster = TRUE,
  plot.height = NULL,
  heatmap.cols = NULL
)
}
\arguments{
\item{heatmap.cols}{}
}
\description{
Title
}
