% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scITDGPlot.R
\name{scITDGPlot}
\alias{scITDGPlot}
\title{Title}
\usage{
scITDGPlot(
  exp.cur,
  row.dist = row.dist,
  k.num = 4,
  time.points = time.points,
  sample.ncell.use = sample.ncell.use,
  cluster.order.list = NULL,
  show.trajectory = TRUE,
  show.term = TRUE,
  OrgDb = org.Hs.eg.db,
  save.wd = NULL,
  use_raster = TRUE,
  plot.height = NULL,
  heatmap.cols = NULL
)
}
\arguments{
\item{heatmap.cols}{}
}
\description{
Title
}
